chore: convert contrib to composable
