+++ components/PoisList/PoisTable.vue
@@ -4,13 +4,13 @@ import { defineNuxtComponent } from '#app'
 import Field from '~/components/Fields/Field.vue'
 import type { ApiPois, FieldsListItem } from '~/lib/apiPois'
 import { PropertyTranslationsContextEnum } from '~/plugins/property-translations'
-import ContributionMixin from '~/mixins/contribution'
+import ContribFieldGroup from '~/components/Fields/ContribFieldGroup.vue'
 
 export default defineNuxtComponent({
   components: {
+    ContribFieldGroup,
     Field,
   },
-  mixins: [ContributionMixin],
   props: {
     fields: {
       type: Array as PropType<FieldsListItem[]>,
@@ -21,7 +21,14 @@ export default defineNuxtComponent({
       required: true,
     },
   },
-
+  data() {
+    const { contribMode, isContribEligible, getContributorFields } = useContrib()
+    return {
+      contribMode,
+      isContribEligible,
+      getContributorFields,
+    }
+  },
   computed: {
     headers(): { value: string, text: string }[] {
       const h = this.fields.map(field => ({
@@ -32,13 +39,11 @@ export default defineNuxtComponent({
         ),
       }))
       h.push({ value: '', text: '' })
-
       if (this.contribMode)
         h.push({ value: '', text: this.$t('fields.contrib.heading') })
 
       return h
     },
-
     context(): PropertyTranslationsContextEnum {
       return PropertyTranslationsContextEnum.List
     },
